Lab4 initial push
I was waiting for you to clean up the repo before commiting this. the base turn signal code with just lights flashing took little time i had no hiccups and everything went how i wanted it to. however, when beginning the simple button this was not the case. hours were spent getting a better understanding of the code now i am comfortable with it and uploading an unfinished but good start to the final for lab 4.