Lab 5 P1 and P2
P1: Reads input data from distance sensor and outputs to serial monitor.

P2: Accelerometer reads data and if parameters are met for X and Y, for example the yellow LED in digital pin 5  has a parameter of 5 degrees. For  the green LED is has a parameter of 1 degree. When conditions are met for these threshold, the results will be an Output at HIGH illuminating the LEDS.