Lab6 Design Challenge
The code initializes a range of sensors and modules, including a Real-Time Clock (RTC), an SD card module, and an accelerometer. Its primary function is to continuously read data from a distance sensor and the accelerometer. When receiving a specific character (‘R’) over the serial port, the code triggers the data collection process. It captures the current timestamp from the RTC and the accelerometer data. These values are directed to a string along with the distance sensor reading. This data string is both written to a CSV file named “LAB6.csv” on the SD card and printed to the serial monitor. The code operates in a loop, repeatedly collecting and logging data, with a delay of 200 milliseconds between each iteration.